backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic-settings
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
pgvector
opencv-python-headless==4.8.1.78
numpy==1.26.4
insightface==0.7.3
onnxruntime-gpu==1.17.1
Pillow
